Return array of created constraints from -autoCreateConstraintsWithoutInstalling:
Enhance the -autoCreateConstraintsWithoutInstalling: API on UIView/NSView to return an array of all the constraints that were created (but not installed) from calls to the PureLayout API inside of the block.
This is particularly useful since any constraints in the block must be captured and stored somewhere, otherwise they will be immediately released (since they are not retained anywhere else).