Fix constraint exception when distributing views
As of iOS 8.0/8.1 SDK, NSLayoutConstraint now raises an exception when attempting to create or activate a location (position) constraint with a multiplier of zero. This is because the 0 multiplier causes the second item to be lost (y=mx+b, with m=0, becomes y=b). Even though this is a valid use case, it's not currently supported. Filed as rdar://19168380

As a workaround, if the multiplier is very close to (or exactly equal to) 0, set it to the minimum value to prevent the second item in the constraint from being lost.

Fixes #16, #32