Add -[configureForAutoLayout] for initialized view
This new API method allows you to set translatesAutoresizingMaskIntoConstraints to NO with a chainable method call (as self is returned). It is convenient when you want to use a specific factory method or initializer, and configure the view for use with auto layout on the same line.

For example:
UIButton *button = [[UIButton buttonWithType:UIButtonTypeSystem] configureForAutoLayout];

Closes #75