Move ALView class methods to NSLayoutConstraint
The following 4 class methods that currently exist on UIView/NSView
make more sense on NSLayoutConstraint:
* `+[autoCreateAndInstallConstraints:]`
* `+[autoCreateConstraintsWithoutInstalling:]`
* `+[autoSetPriority:forConstraints:]`
* `+[autoSetIdentifier:forConstraints:]`

This change makes the biggest improvement in Swift, when using the
`autoSetPriority` or `autoSetIdentifier` methods with a trailing
closure.

This is a breaking API change, however all of these methods are simply
moving, with no change in behavior.

Additionally, some of the identifier methods were protected by the
MinBaseSDK_iOS_8_0 macro, but weren’t also using the
MinBaseSDK_OSX_10_10 macro, meaning these APIs would not be usable for
a Mac target. This has been corrected.